Move tests to one of file_contents_tests, file_structure_tests, or other_tests
