+++ tests/other_tests.rs
@@ -110,49 +110,3 @@ fn test_create_structure() {
     path.push("data/test_create_structure/functions/main.databind");
     assert!(path.exists() && path.is_file());
 }
-
-/// Test running `databind` alone in a created project.
-/// Runs deeper than project root to ensure that it can both
-/// find the `databind.toml` file and that the output folder
-/// is in the correct directory
-#[test]
-fn test_databind_alone() {
-    let out = TempDir::new("test_databind_alone").expect("Could not create tempdir for test");
-    let mut path = PathBuf::from(out.path());
-
-    // Create project
-    tests::run_with_args(
-        "cargo",
-        &[
-            "run",
-            "--",
-            "create",
-            "test_databind_alone",
-            "--path",
-            out.path().to_str().unwrap(),
-        ],
-        None,
-    );
-
-    path.push("src");
-    println!("running in path: {}", path.display());
-
-    // Run `databind` in directory
-    tests::run_with_args(
-        "cargo",
-        &[
-            "run",
-            "--manifest-path",
-            &format!("{}/Cargo.toml", env!("CARGO_MANIFEST_DIR"))[..],
-        ],
-        Some(&path),
-    );
-
-    // Pop /src from the path to check for /out in the project root
-    path.pop();
-
-    // Test that out directory exists
-    path.push("out");
-    println!("Checking path {}", path.display());
-    assert!(path.exists() && path.is_dir());
-}
